backend/requirements.txt
fastapi
uvicorn
langgraph
qdrant-client
sentence-transformers
google-generativeai
pydantic
python-dotenv
requests
